requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
ultralytics>=8.3.0
opencv-python-headless==4.8.1.78
python-multipart==0.0.6
pydantic==2.5.0
numpy==1.24.3
python-dotenv==1.0.0
